third_party/blink/web_tests/platform/generic/fast/table/row-in-inline-block-expected.txt

Test that a table-row child of an inline is wrapped inside an inline-table (and table-row-group).

There should be one blue box above and one below this line.
PASS lastTop is > firstBottom
PASS successfullyParsed is true

TEST COMPLETE
